refactor: WinningNumber으로 변경
