+++ README.md
@@ -1,4 +1,4 @@
-# java-lotto-kakao
+# java-lottoNumber-kakao
 
 ## 기능 요구사항
  
@@ -6,7 +6,7 @@
 * 구매 금액만큼 로또를 무작위로 생성한다.
 * 생성된 로또를 출력한다.
 * 지난 주 당첨 번호 + 보너스 볼을 입력한다.
-* 당첨 번호와 비교해 로또의 등수를 계산한다.
+* [x] 당첨 번호와 비교해 로또의 등수를 계산한다.
 * 당첨 결과를 출력한다.
 * 수익률은 소수 둘째자리까지만 출력한다.
